services/office/pyproject.toml
[project]
name = "briefly-office"
version = "0.1.0"
description = "Briefly Office Integration Service"
requires-python = ">=3.12"
dependencies = [
    # Web Framework
    "fastapi",
    "uvicorn[standard]",
    "pydantic>=2.0.0,<2.9.0",
    "python-dotenv",
    "structlog",
    "httpx",
    "requests",
    "email-validator",

    # Database
    "sqlmodel>=0.0.21",
    "sqlalchemy>=2.0.0",
    "aiosqlite",
    "asyncpg",
    "alembic",

    # External API Integrations
    "google-auth",
    "google-auth-oauthlib",
    "google-auth-httplib2",
    "google-api-python-client",
    "msal",
    "PyJWT",

    # Caching
    "redis",

    # GCP
    "google-cloud-secret-manager",

    # OpenTelemetry
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-instrumentation",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-instrumentation-httpx",
    "opentelemetry-distro",
    "opentelemetry-exporter-gcp-trace"
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
